更新全局消息组件
