+++ src/service/Service.js
@@ -5,8 +5,8 @@ export default class Service extends BaseService {
    * 用户登录
    * **/
   static login () {
-    Service.get('/api/v1/login').then((token) => {
-      console.log(token)
-    })
+    // Service.get('/api/v1/login').then((token) => {
+    //   console.log(token)
+    // })
   }
 }
